add generic methods to create derived socket store
